// components/admin/adminClearAttempts/adminClearAttemptsActionButtons/adminClearAttemptsActionButtonsStyle.scss.d.ts
export type Styles = {
  'btnContainer': string;
  'btnContainerItem': string;
  'btnContainerItemDescription': string;
  'buttonShowDetailUserInfo': string;
  'buttonShowDetailUserInfoClose': string;
  'buttonShowDetailUserInfoContainer': string;
  'buttonShowDetailUserInfoOpen': string;
  'componentContainer': string;
  'contentContainer': string;
  'dropdownContainer': string;
  'dropdownTestAndUserContainer': string;
  'editorWrapper': string;
  'iconBtn': string;
  'isActiveContainer': string;
  'pulse': string;
  'showHideContainer': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;